- Few modification completed

+++ app/controllers/sessions_controller.rb
@@ -13,7 +13,11 @@ class SessionsController < Devise::SessionsController
     sign_in(scope, resource) unless warden.user(scope) == resource
     respond_to do |format|
       @success = true
-      format.html {redirect_to session.has_key?(:return_url) ? session.delete(:return_url) : root_path}
+      if current_user.role? :super_admin
+        format.html {redirect_to admin_dashboard_path}
+      else
+        format.html {redirect_to session.has_key?(:return_url) ? session.delete(:return_url) : root_path}
+      end
     end
   end
